+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -2243,7 +2243,11 @@ function Get-AllSPOPackages
 
         [Parameter()]
         [System.String]
-        $CertificateThumbprint
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [Switch]
+        $ManagedIdentity
     )
 
     try